+++ script/tool/lib/src/license_check_command.dart
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:file/file.dart';
+import 'package:git/git.dart';
 import 'package:path/path.dart' as p;
 
 import 'common/core.dart';
@@ -105,7 +106,8 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 /// Validates that code files have copyright and license blocks.
 class LicenseCheckCommand extends PluginCommand {
   /// Creates a new license check command for [packagesDir].
-  LicenseCheckCommand(Directory packagesDir) : super(packagesDir);
+  LicenseCheckCommand(Directory packagesDir, {GitDir? gitDir})
+      : super(packagesDir, gitDir: gitDir);
 
   @override
   final String name = 'license-check';
@@ -117,9 +119,16 @@ class LicenseCheckCommand extends PluginCommand {
   @override
   Future<void> run() async {
     final Iterable<File> allFiles = await _getAllFiles();
+    // Create a set of absolute paths to submodule directories, with trailing
+    // separator, to do prefix matching with to test directory inclusion.
+    final Iterable<String> submodulePaths = (await _getSubmoduleDirectories())
+        .map(
+            (Directory dir) => '${dir.absolute.path}${platform.pathSeparator}');
 
     final Iterable<File> codeFiles = allFiles.where((File file) =>
         _codeFileExtensions.contains(p.extension(file.path)) &&
+        !submodulePaths.any(
+            (String submodule) => file.absolute.path.startsWith(submodule)) &&
         !_shouldIgnoreFile(file));
     final Iterable<File> firstPartyLicenseFiles = allFiles.where((File file) =>
         path.basename(file.basename) == 'LICENSE' && !_isThirdParty(file));
@@ -275,6 +284,24 @@ class LicenseCheckCommand extends PluginCommand {
       .where((FileSystemEntity entity) => entity is File)
       .map((FileSystemEntity file) => file as File)
       .toList();
+
+  // Returns the directories containing mapped submodules, if any.
+  Future<Iterable<Directory>> _getSubmoduleDirectories() async {
+    final List<Directory> submodulePaths = <Directory>[];
+    final Directory repoRoot =
+        packagesDir.fileSystem.directory((await gitDir).path);
+    final File submoduleSpec = repoRoot.childFile('.gitmodules');
+    if (submoduleSpec.existsSync()) {
+      final RegExp pathLine = RegExp(r'path\s*=\s*(.*)');
+      for (final String line in submoduleSpec.readAsLinesSync()) {
+        final RegExpMatch? match = pathLine.firstMatch(line);
+        if (match != null) {
+          submodulePaths.add(repoRoot.childDirectory(match.group(1)!.trim()));
+        }
+      }
+    }
+    return submodulePaths;
+  }
 }
 
 enum _LicenseFailureType { incorrectFirstParty, unknownThirdParty }