+++ script/tool/test/license_check_command_test.dart
@@ -7,12 +7,14 @@ import 'package:file/file.dart';
 import 'package:file/memory.dart';
 import 'package:flutter_plugin_tools/src/common/core.dart';
 import 'package:flutter_plugin_tools/src/license_check_command.dart';
+import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
 
+import 'common/plugin_command_test.mocks.dart';
 import 'util.dart';
 
 void main() {
-  group('$LicenseCheckCommand', () {
+  group('LicenseCheckCommand', () {
     late CommandRunner<void> runner;
     late FileSystem fileSystem;
     late Directory root;
@@ -23,8 +25,12 @@ void main() {
           fileSystem.currentDirectory.childDirectory('packages');
       root = packagesDir.parent;
 
+      final MockGitDir gitDir = MockGitDir();
+      when(gitDir.path).thenReturn(packagesDir.parent.path);
+
       final LicenseCheckCommand command = LicenseCheckCommand(
         packagesDir,
+        gitDir: gitDir,
       );
       runner =
           CommandRunner<void>('license_test', 'Test for $LicenseCheckCommand');
@@ -121,6 +127,25 @@ void main() {
       for (final String name in ignoredFiles) {
         expect(output, isNot(contains('Checking $name')));
       }
+    });
+
+    test('ignores submodules', () async {
+      const String submoduleName = 'a_submodule';
+
+      final File submoduleSpec = root.childFile('.gitmodules');
+      submoduleSpec.writeAsStringSync('''
+[submodule "$submoduleName"]
+  path = $submoduleName
+  url = https://github.com/foo/$submoduleName
+''');
+
+      const String submoduleSourceFile = '$submoduleName/foo.dart';
+      root.childFile(submoduleSourceFile).createSync(recursive: true);
+
+      final List<String> output =
+          await runCapturingPrint(runner, <String>['license-check']);
+
+      expect(output, isNot(contains('Checking $submoduleSourceFile')));
     });
 
     test('passes if all checked files have license blocks', () async {